refactor: optimise template provider
