--- projects/integration/src/app/ui-theme.component.ts
@@ -1,11 +1,0 @@
-import {Provider} from "@antischematic/angular-state-library";
-import {Directive, ElementRef, inject} from "@angular/core";
-
-@Directive({
-   standalone: true,
-   selector: "Theme",
-})
-export class UITheme extends Provider {
-   color = "red"
-}
-
