+++ projects/integration/src/app/ui-theme.ts
@@ -1,0 +1,10 @@
+import {TemplateProvider} from "@antischematic/angular-state-library";
+import {Directive} from "@angular/core";
+
+@Directive({
+   standalone: true,
+   selector: "ui-theme"
+})
+export class UITheme extends TemplateProvider {
+   color = "red"
+}
