+++ projects/integration/src/app/ui-todos.component.html
@@ -35,6 +35,8 @@
    Toggle All
 </button>
 
+<button (click)="shuffle()">Shuffle</button>
+
 <style>
    :host {
       display: block
